dried simple rpc methods
